services/auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary
asyncpg
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic
pydantic-settings==2.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-fastapi-instrumentator
